+++ com.unity.cinemachine/CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [2.9.8] - unreleased
 - Bugfix: Occasional precision issue when camera rotation is exactly 180 degress, causing roitational flickering.
+- Bugfix: CinemachineCollider was causing a pop when OnTargetObjectWarped was called.
 
 
 ## [2.9.7] - 2023-05-06